requirements.txt
Flask-SQLAlchemy
Flask-Security
Flask-Login
flask-discoverer
requests
alembic
sqlalchemy-utils
